1_Q standard RB without and with GSP curve tail
The comparison of the results of the two notebooks shows that obtaining the curve tail of the ground state population curve is mandatory. Increasing the maximal circuit length results in a B value nearer the ideal expected one (0.5) and has an effect on the calculated error per Clifford.
In addition, the second notebook shows the limits of the hierarchical method (see the warning after running SMC).